@@ -1,7 +1,13 @@
 {
     "name": "laravel-zero/laravel-zero",
     "description": "The Laravel Zero Framework.",
-    "keywords": ["framework", "laravel", "laravel zero", "console", "cli"],
+    "keywords": [
+        "framework",
+        "laravel",
+        "laravel zero",
+        "console",
+        "cli"
+    ],
     "homepage": "https://laravel-zero.com",
     "type": "project",
     "license": "MIT",
@@ -16,7 +22,7 @@
         }
     ],
     "require": {
-        "php": "^7.3",
+        "php": "^7.4",
         "laravel-zero/framework": "^8.0"
     },
     "require-dev": {
@@ -41,5 +47,7 @@
     },
     "minimum-stability": "dev",
     "prefer-stable": true,
-    "bin": ["cart"]
+    "bin": [
+        "cart"
+    ]
 }
